Add schema synchronize step to module update process
For simple changes, adding tables, columns, and indexes or changing column attributes the changes are automatic during module update, and no additional scripting is required. For some cases, for example a column rename, the schema synchonizer cannot reliable distinquish between a rename or a drop of one column and an add of another. For these cases the preupdate script hook can be used to alter the table.

If defined, the onUpdate xoops_module_preupdate_(dirname) script  is called immediatly before the schema sync. This provides the opportunity to make any major database changes which might destroy data if left to the schema sync alone, such as table and column renames.