Implement support for travis-ci
Includes a console command to allow minimal Xoops install from command line, and .scurinizer.yml and .travis.yml updates to run ci chain on git push.

As is, the console lives in the console directory, and can run from a fresh clone without additional configuration once a 'composer install' is run. It is kind of held together with duct tape, but it runs. Hopefuly, it demonstrates the value of the techinique and will serve as a base for future development.