Set a specific namespace on cache pools
Since the cache and database are separate, it is possible to create a circumstance where the system represented in the database and the system represented in the cached entities are not the same system. By setting the namespace based on the database prefix of the system generating the entries, the effect of an accidental mismatch is minimized.

An example of the mismatch would be a module installed according to the cached list of installed modules, while it is not installed accoring to the  database.