Switch to patchwork/jsqueeze for javascript minimizing
This is **much** faster than JSMin -- especially noticeable on first entry to the admin area.

NB - requires composer update, as well as a change to xoops_data/configs/system_assets_pref.yml to replace 'jsmin' with 'jsqueeze'.
Easiest to just delete the file if you have not customized it, and it will be rebuilt. A fresh install will automatically use this as the default.